tdf#130857 qt weld: Support checkbox inside button box
QCheckBox inherits from QAbstractButton, so can be
placed inside of a QDialogButtonBox.

Call QtBuilder::setButtonProperties from
QtBuilder::setCheckButtonProperties to make that
case work as expected, as the former takes care
of inserting the QAbstractButton into the button box.
It also takes care of handling the "label" property.

This is used e.g. by the Welcome dialog
(cui/uiconfig/ui/welcomedialog.ui) and fixes the
checkbox text being cut off when using the qt6 VCL
plugin with SAL_VCL_QT_USE_WELDED_WIDGETS=1 for the
following scenario:

* copy over profile from older LO version, e.g.
  copy instdir/user from LO 25.2 bibisect repo
* start Writer

With this commit in place, the checkbox text is no
long cut off. The checkbox is shown at the end
instead of at the beginning now, however, which will
be addressed in a separate commit.

Change-Id: Ibaa955798c87db2a41760dd0e6106238ddeff850
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186444
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins