tdf#130857 cui: Set proper position for checkbox in Welcome dialog
Set the "Do Show again" checkbox in the Welcome dialog
to be at position 0 to make sure this is the first item
in the button box. While gtk3 and gen show the checkbox
at that position even when its position is set to the last one
in the .ui file (i.e. without this change in place), that's
not the case for qt6 with SAL_VCL_QT_USE_WELDED_WIDGETS=1.

This fixes the following issue mentioned in previous commit

    Change-Id: Ibaa955798c87db2a41760dd0e6106238ddeff850
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jun 13 09:54:29 2025 +0200

        tdf#130857 qt weld: Support checkbox inside button box

:

> With this commit in place, the checkbox text is no
> long cut off. The checkbox is shown at the end
> instead of at the beginning now, however, which will
> be addressed in a separate commit.

Change-Id: Ie2a9c493920bc835a99f434974d94ff744b834c1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186445
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>