Stop using UNO abstraction for color picker
SvColorDialog is using an instance of
ColorPickerDialog under the hood.

So far, the ColorPickerDialog instance was
created and used via UNO abstraction, by
calling either
css::cui::ColorPicker::createWithParent (for the
synchronous case) or
css::cui::AsynchronousColorPicker::createWithParent
(for the async case), which would end up calling
the UNO service ctor,
com_sun_star_cui_ColorPicker_get_implementation
(and creating an instance of ColorPickerDialog
in both cases, just with a reference to a different
UNO interface as a return value for both cases).

Further interaction would then happen via the
XExecutableDialog or XAsynchronousExecutableDialog
UNO interfaces and by setting/reading properties.

Stop using that UNO abstraction and instead
introduce AbstractColorPickerDialog as a new
VclAbstractDialog subclass, which allows to
interact with the the ColorPickerDialog more
directly while still avoiding a direct dependency
on (the higher-level module) cui in svtools.

This also implies that SvColorDialog now no longer
depends on svt::DialogClosedListener for the async
case, which means that SvColorDialog can be moved
from svtools to vcl in the future. That in turn will
help in implementing support for platform/toolkit-native
alternatives for the color picker instead of using
the LO-internal one.

No change in user-observable behavior intended
or seen in a quick test in Writer (open dropdown for
"Font Color" toolbar item, then select "Custom Color"
to open the dialog).

Change-Id: I3b9997b73a9a6bcce4c5b3ae3bd7d57816a4419b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186468
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>