tdf#144743 Allow guide name with starting digit
In OOXML, it is allowed for a guide name to start with a digit. The
presetShapeDefinitions funnel, mathEqual, and mathNotEqual have such
guides.

If a formula uses such a guide, LO imported it as a number instead of
a reference to the guide. The problem occurs with both prstGeom and
custGeom.

Change-Id: I7cda3be05d98c3101e84da0dcb01686a97baa77f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186478
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>