Related: tdf#167018 Generalize and deduplicate Theme::FromAny
Let it also handle a hypothetical non-UnoTheme implementation (it
is possible to create these e.g. from Basic / Python code). When
handling it, use theme name as color set name (there seems to be
no UNO API to get color set name?). And use the same code for the
creation of Theme from PropertyValue list.

The check that enumrange<ThemeColorType> does not iterate through
ThemeColorType::Unknown is redundant, since that value is outside
of [0, LAST] range.

Change-Id: Iae6770f196011deeb2a92d3b64af514fb6a53d06
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186513
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins