Use vcl::ColorPickerMode instead of sal_Int16 param
Now that the UNO abstraction and corresponding services
have been dropped in

    commit 0905c23676fe34a5346683ee46568e4d1c58806a
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jun 13 18:27:09 2025 +0200

        Stop using UNO abstraction for color picker

and

    Change-Id: I9264707df57e46192a393f6736257c413e5d9831
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Jun 16 08:01:55 2025 +0200

        [API CHANGE] Drop ColorPicker, AsynchronousColorPicker

, no longer pass a sal_Int16 param to the ColorPickerDialog
ctor, but use the vcl::ColorPickerMode enum directly
without casting.

This also means that there is no more need for the
underlying int values of the enum values to have a
specific value, so no longer set these explicitly.

Change-Id: Ife2a7b0672bacc73b3001ef0e2e57f4626837655
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186539
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins