tdf#167033: allow double processing of Writer shapes' text frame nodes
It turns out, that these are naturally processed twice.

When enumerating paragraph's content, CollectFrameAtNode would find both
the shape and the shape's text frame anchored to the paragraph node.
Then, XMLTextParagraphExport::exportTextContentEnumeration will process
the shape; and in XMLShapeExport::collectShapeAutoStyles, the shape will
be queried for XText, which will return the text frame, which will then
be processed.
And finally, XMLTextParagraphExport::exportTextContentEnumeration will
process the text frame, for the second time - at which point, the frame's
paragraphs will be processed again.

I don't know if it's possible to avoid the repeated processing; here I
only add a debug check for this, and skip adding duplicate node index.

Change-Id: I84275f233554a6b4923ddaabc336bec7a94ec2fa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186546
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>