tdf#165742 Step 4.6: Establish a narrow export path for chartex
This is a subtask of tdf#165742: Chartex charts are lost on input from
OOXML and re-export.

Fix an existing bug in chart import, which results in the incorrect
default value for stroke width on chart shapes. This was discovered
in testing chartex and can be reproduced in a hand-edited pre-2016
file (<c> namespace), but may not manifest there in practice since
it appears MS Office may always output an explicit stroke width for
pre-2016 charts.

Change-Id: I458ae6ad0eaf28bf933c8ae0baed36e48df114b7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186582
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins
(cherry picked from commit c8b26350d9aaf9cc8f8202ab66d199700f3be604)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187398
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>