tdf#167044 sw RTF import: fix bad image size in shape text
Regression from commit 57d9cc81b058757421cd082e5dbe32a919716807
(tdf#127806: RTF import: use shape dimensions for pib picture,
2022-09-01), open the bugdoc, the icon image inside the shape should be
about 0.37cm x 0.37cm, is about 14.98 cm x 12.66 cm.

What happens is that \pict was assumed to be inside the property value
of a shape, in which case both the shape and the picture can have a
size, and the old use-case requires to ignore the inner size. The new
use-case is two user-visible shapes: an outer text frame and an inner
small icon image, inside shape text. The different sizes are wanted in
this case.

Fix the problem by conditionally going back to the old behavior of not
inheriting size for the inner shape. The name of the destination at the
top of the parser stack is different for the "in property value" and the
"in shape text" cases, so this keeps the old use-case working while
restores the wanted small size for the new use-case.

Here the difference between the old and new size was so large that the
end-user saw this as "the image was not visible previously".

Change-Id: Ife93089b967fc26b7180775db8fc507a896d160b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186602
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit e0e415ba89925a933f8a07e4a2524269bd5c7495)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186627
(cherry picked from commit d81328b59856d2b370110dd1d92a9cce27af34f0)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186642
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>