tdf#119192 editeng: Fix for copy-and-paste using font from wrong script
To minimize the size of the output RTF, Edit Engine only emits font
attributes when either the font changes or when the script type changes.

In the latter case, Edit Engine was previously writing font attributes
inside a child scope, rather than the paragraph scope. Font changes
related to a language change would thus apply only to the first run of
characters in that language, while following runs would appear to use an
arbitrary language.

Change-Id: Iff38299bffaf3045cef5c6e71b8cab441ba0be39
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186612
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>
Tested-by: Jenkins