Do not drop read-only flag when searching for a slot
bReadOnly reflects the environment state. During the lookup, for some
slots, this flag could be reset, because that slot was eligible for
an exception from general rule. Yet, there could be a situation, when
a slot has changed the flag, but then was not picked (see 'continue'
in the end). Then this modified flag state would be used for the next
slot, which was wrong.

Change-Id: Ie2f7976de6314f4f50b187c286d189f875db527e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186621
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>