sw: Navigator: fix SwPostItContent::m_pField UAF
ASAN reports UAF on SwPostItContent::m_pField and it's apparent that the
only thing that removes a SwPostItContent from the Navigator is a timer.

Clearly it becomes invalid when the SwFromatField that m_pField points
to is destroyed, so implement a listener for this and try to add some
null checks on uses to avoid crashing. (Doesn't look obvious how to
remove it from the tree immediately.)

Change-Id: Iad94ddb923f0126b5e6d76d0b77dc26043cefaba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186654
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins