tdf#166319 sw interdependent redlines: fix redo of reject of del-then-fmt's del
The bugdoc has <del>AA<format>BB</format>CC</del>, rejecting AA just
leaves the BB format redline, undo works, but redo results in no
redlines, while we again want a single format redline.

The first problem was that the del-then-fmt redline's undo action was an
SwUndoAcceptRedline, while the UI action was a reject. This was not a
problem on undo, since both accept and reject has the undo code to
restore a previous state, but redo will have to perform different
actions (re-accept or re-reject).

The second problem was that sw::DocumentRedlineManager::RejectRedline()
assumed that rejecting an underlying redline (depth = 1, instead of the
usual depth = 0) could be always implemented by accepting the redline,
since that works for the insert/delete combination. Improve this, so in
case we get a format-on-delete, then we again properly reject just the
underlying redline, as the original UI action did.

Change-Id: I846abcfa34a7d55709bc9af6e566a3e83d79d833
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186700
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>