LOK: Introduce "read-only mode with tracked changes management"
When it is enabled, most of the editing commands in a read-only view
are disabled, as usual; but the commands to manage tracked changes
are enabled, and allow user to accept / reject / comment on changes.
Saving is enabled, too.

Change-Id: Ic8bafc335c47ecd1b897c428a804cea2bc677cb3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186686
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit c4f3d75567d5664259e01ff14a75e6f3e952ac9a)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186733
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>