a11y tests: Introduce C++ XAccessibleTextTester
Similar to how e.g. XAccessibleContextTester tests
the methods of the XAccessibleContext interface, this
one tests the methods of the XAccessibleText interface.

The logic in this C++ version is in general based on the Java
implementation in
qadevOOo/tests/java/ifc/accessibility/_XAccessibleText.java.

This will be used to port the AccessibleFixedTest
Java test to C++ one in an upcoming commit.

Tests that don't test common expectations for all XAccessibleText
implementations but depend on further conditions (like particular
text content, see the tests in the Java version that check whether
or not this is the "SmGraphicAccessible" or "AccessibleStatusBarItem"
test) are not included in the C++ version.
If considered relevant, I think it makes more sense to include those
specific tests in the test suite for the particular classes for which
the corresponding behavior is expected.

In XAccessibleTextTester::testGetCharacterBounds, don't
fail if the character bounds are outside of the
component, but just produce output, as the Java
version (in _XAccessibleText._getSelectedText) does.

Change-Id: Iec0c6a7413181b2f6ea28fdbbf6c0f2d9792fac8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186677
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins