a11y tests: Allow custom function to generate a11y events
So far, the XAccessibleEventBroadcasterTester logic was
to modify a vcl::Window to generate a11y events for it
that the XAccessibleEventBroadcaster would receive.

Change this to allow passing a function to the ctor that
is responsible for generating accessible events, to
allow using this class for other scenarios in the future as
well.

Move the vcl::Window-specific logic to a new subclass
WindowXAccessibleEventBroadcaster.

Change-Id: I36cd49d40e68127d2d98a6b5c151b1e32c29dd92
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186678
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins