Emscripten --disable-gui: Improve waiting yield
This call to ImplYield (with bWait = false) had been added in
858fed2a98ad5a04ca65dc504722e2db52746e64 "Emscripten: Support --disable-gui",
and calling it with bWait = false ensures that e.g.
<https://github.com/allotropia/zetajs/tree/main/examples/convertpdf>, when
running against a --disable-gui LO build, does not deadlock.  But for use cases
involving LOK, it instead needs to be called with bWait = true, as does the
normal Application::Yield in vcl/source/app/svapp.cxx:  Otherwise, the code at
<https://github.com/CollaboraOnline/online/tree/master/wasm> would never proceed
past the

>     if (!bMustSleep && !comphelper::LibreOfficeKit::anyInput())
>         return bWasEvent;

in SvpSalInstance::ImplYield.

Change-Id: I985a75bea042b3191d95fef96a5430faeb6a521e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186665
Reviewed-by: Stephan Bergmann <stephan.bergmann@collabora.com>
Tested-by: Jenkins