java a11y tests: Don't rely on XWindow being XAccessible
In order to retrieve the XAccessible associated with
a window in AccessibilityTools.getAccessibleObject,
no longer directly query the XWindow for the XAccessible
interface, but instead (temporarily) introduce and use a
new "XAccessible" property for the deprecated XVclWindowPeer
interface (which "gives access to the VCL window implementation")
in VCLXWindow to get the XAccessible for the underlying
vcl::Window.

This ensures this will continue to work when the VCLXWindow
no longer is the accessible for its associated vcl::Window,
s. upcoming commit

    Change-Id: I25de1ab0f8b47b12346abd01bad05cb54fb75f52
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 26 09:45:11 2025 +0200

        vcl a11y: Let vcl::Window create accessible directly, drop a11y factory

Once all Java a11y tests have been ported to C++, this
property will no longer be needed, because direct access to
the vcl::Window will be possible then, see e.g.

    commit d3262b72f69311b282fbdbe3c38de743e76bdda9
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 26 09:23:29 2025 +0200

        a11y tests: Don't rely on XWindow being window's XAccessible

, but Java tests rely on the information being available
somehow via UNO API.

Change-Id: Id96a2104d87c998b0bdcbce36451499484d81272
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186680
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins