tdf#167128: SwRedlineAcceptDlg: move enabling controls after selection
It was done in InitAuthors, which was called before lcl_reselect;
and therefore, opening the dialog would not see the selection, and
not activate buttons properly.

Now the "Accept" and "Reject" are active upon dialog opening, when
there is a selection.

This required extracting the respective code to a separate function.

Change-Id: I3a02384b402312613fef4b2d3932d19c9b1acc19
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186752
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186767