a11y: Drop "=====  internal  ======" comments
It's unclear to me what these

    //====  internal  =========================================================

comments are trying to say.

(What does it mean if the  public ctor of a class that
gets called from outside that class is "internal"?)

If those were intended to say what methods are internal to a class
then access specifiers (private, protected, public) would be the proper
way to specify what is internal to a class (but that doesn't seem to
be the semantic used here).

If "internal" is meant to indicate that these are LibreOffice-internal
(not available for third-party code in extensions or
macros), then there is no need to explicitly mention that, because
all a11y class implementations are internal only, no instances are
created directly by calling ctors from third party code.)

Change-Id: I118c7e0bee77cda68e67f0a3bfe134aeaaa86d73
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186746
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>