cool: reduce time spent in FillInfo
when rendering large sheets, we spend a lot of time traversing the surrogates
because, for complex documents, we end up with a lot of surrogates that
are unrelated to the one we are actually interested in.

This was probably done as a performance optimisation, and is probably a good
idea for very specific documents ie. ones that ONLY have a single kind of surrogates.

It makes much less sense for more general documents. Also, we are trying to get away from surrogates,
so drop ATTR_ROTATE_VALUE as a surrogate, and rely on scanning the ScAttrArray.

This reduces the time spent scanning surrogates in favour of scanning
ScAttrArray, which improves performance for more complex documents.

Change-Id: I8a1cb634185065ba4d6b9b7682fcd9e2a33fe965
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187014
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>