sc: initial commit of "sheet view" functionality
The idea of a sheet view is that we can change the auto filter's
fitlering and sorting in the current view without influencing
the filter in other views.

This is realised with a copy of a current sheet, which is shown
in the current view instead of the default sheet (which still
can be seen by other views). Very important aspect is to keep the
tables in sync.

There can be multiple sheet views and in the current view we can
switch between the default view and other sheet views freely.

This change adds a command to create a new sheet view. When the
command is triggered, it then creates a copy of the current sheet
and registers a new sheet view for that sheet/table. The current
view automatically uses the copy of the sheet for all changes, but
if the sheet is changed, nothing is synched yet.

Change-Id: Ia77c5ab759113eaae237c638015401418ab56033
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189326
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Kohei Yoshida <kohei@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>