tdf#166978: replace HUGE_POSITIVE hack with proper upper invalidation
The problem was, that the change of section's height (either to zero
when hiding, or from zero to some height when showing) could keep its
upper's size unchanged and valid. Invalidating the upper's size in
SfxHintId::SwSectionHidden handler was not enough; in the following
layout calculation, the upper sets its mbFrameAreaSizeValid to true
before its lowers' calculation. Then, during their calculation:

* If the section was hiding, SwFrame::MakeValidZeroHeight would fail
  in ShrinkFrame, then force its height to 0 in 'if (IsLayoutFrame())'
  block, and not notify upper;
* If the section was showing, SwSectionFrame::Grow_ could see that
  the upper has space enough for the section (e.g., when there are
  more children after this section), and also not notify upper.

The HUGE_POSITIVE hack allowed to force following re-layout in some
cases; but the bugdoc failed when the cell was initially made very
tall, and hat to split; then the section in it was shrinked to 0,and
that didn't cause the merge with the follow.

This fixes the problem by:
1. Removing of the HUGE_POSITIVE hack;
2. Invalidating upper cell frame in SwSectionFrame::Grow_, similar
   to commit ee57229154a91cf632871cfcb1eb9609908f3882 (Fix #88178#:
   Sectionframe in headerframe, don't grow without invalidation,
   2001-06-13);
3. Invalidating layout frame's upper in SwFrame::MakeValidZeroHeight,
   when ShrinkFrame didn't make it 0-height.

In #3, I decided to not check the upper's kind, for robustness; if
needed, checks may be added in follow-ups, for specific problems.

Change-Id: Iafdd1c41b52d2170d5dd2b3bc2cbc1a5f657c0fb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186779
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>