tdf#121493: handle \trleftN after \cellxN
I took an easy path, and just fixed the possible pre-existing value
when \trleftN is handled. But the proper fix would be to avoid the
cell widths calculations altogether, until \row is handled. At that
point, all information is already gathered.

Search for the first instance of LN_CT_TblGridBase_gridCol, that is
not -1 (the special value inserted in RTFDocumentImpl::resetTableRowProperties,
and used in DomainMapperTableManager::sprm); and if found, correct
it by the calculated TRLeft value.
When there was no previous \cellxN, correct rCurrentCellX.

Change-Id: Ie6cf64d594e166abb1ed3e939207101d96f75c63
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186790
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit a06689b1711e1f200afd4a4492341f337ba99082)
Reviewed-on: https://gerrit.collaboraoffice.com/c/core/+/1373
Reviewed-by: Thorsten Behrens <thorsten.behrens@collabora.com>
Tested-by: Thorsten Behrens <thorsten.behrens@collabora.com>