lok: CellsAtX, CellsAtY can return wrong values
In lok case the document on the core side is never scrolled, so the
document position is always the top of the document (that is (0,0))
That means that the screen width and height can be greater than
sal_uInt16 max value.
So, in order to avoid unexpected truncation we need to use tools::Long
in place of sal_uInt16.
In fact before this patch we missed to report the correct CellsAtY value
for cells below row 10000 and the correct CellsAtX value for cells
beyond column AZZ.
On cell editing that caused edit area width or height to be not expanded
properly.

Change-Id: I9f38b55ee9b1edbf9b5f0717155597f363bf9768
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186799
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
(cherry picked from commit bfa9d2aacd5188761cbde343748b5f6cd2c12583)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186916
Reviewed-by: Marco Cecchetti <marco.cecchetti@collabora.com>
Tested-by: Jenkins