tdf#167169: postpone \trleftN processing until \row
... where all the data: all \cellxN and \tblindN - are already known.
The \tblind establishes the left indent; \trleft establishes what X
coordinate will be at the leftmost table edge (i.e., \tblindN tells
"shift table N twips to the right"; \trleftN tells "assume that the
coordinate of this table's leftmost position is X"). Only when there
is no \tblindN, the \trleftN also defines an indent. Handling all of
that correctly is only possible at \row handler; especially because
the RTF standard tells:

  <row> (<tbldef> <cell>+ <tbldef> \row) | (<tbldef> <cell>+ \row) | (<cell>+ <tbldef> \row)

  ...

  While Word 97 emitted the row properties (<tbldef>) at the beginning
  of the row, a reader should not assume that this is the case.
  Properties can be emitted at the end, and, in fact, Word 2002, Word
  2003, and Word 2007 do this.

Note: there is a problem of inconsistencies in the documentation and
implementation of \tblindN wrt. units: there is \tblindtypeN, which
should define units used in \tblindN, and it is documented to use 0
for "auto", 1 for "dxa" (=twips), 2 for "no width", and 3 for "pct"
(fiftieths of a percent). However, in my testing, Word considered 3
as twips, and ignored \tblindN when \tblindtypeN had other values.
In Writer, \tblindtypeN is not handled. This needs a separate fix.

Change-Id: If5c47571ca005b248d72ac7ab688c358f4f18e52
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186807
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186939