Chart: implemented chart themes (experimental)
This commit content is hided behind ExperimentalMode.

Added a sidebar panel to set a chart style.
Similar to MS office "chart style".

It sets many chart properties at once, but now mostly font related.
( It is planned to extend it with many other properties )
We can even save a chart style into the themes there with a button.

Added a dialog to manage chart themes.
-we can save the selected chart properties into a new theme
-we can overwrite a theme with the selected chart.
-we can set theme properties to the selected chrat.
-we can delete a theme
( MSO does not able to do these. :) )

there is 4 pre-defined theme that user cannot change.
( The 4 pre defeined chart theme is only a test case placeholder
it is planned to make them more sensible and more of them. )

There is some todo with it:
-render of the chart thumnbail still not fast enought.
(now, ChartView::update() is the slow)
-we could probably find better places for the predefined themes
-expand the saved properties with many more.
(already added the legends position)
-export/import of the user generated themes

Change-Id: Ie93a2561df92c54cea1746d59dec665cda929060
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186950
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/198495
Reviewed-by: Marco Cecchetti <marco.cecchetti@collabora.com>