tdf#166092 writerfilter: exception killed critical settings
I also took the opportunity to move one setting
(that was placed directly into the DTOR)
into where these things are normally stored in ApplyProperties
(which is called just a bit earlier in that same DTOR).

I tried setting an assert when this fails,
but already hit
sw/qa/writerfilter/filters-test/data/pass/CVE-2005-2964-1.rtf

An earlier patchset contains the unit test.
However, I did not include it because it takes a VERY long time
to load in a debug environment,
and it seems pointless to try to create a minimal document
that still causes a very strange footnote exception.

Change-Id: Ia2b38232e21149407bccceea4ac859d1f38426e6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186856
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>