tdf#166141 writerfilter docx: linkStyles != 0.35cm bottom margin
17.15.1.55 linkStyles
(Automatically Update Styles From Document Template)
This element specifies that styles in the given document
shall be updated to match the styles in the attached
template specified using the attachedTemplate element (§17.15.1.6)
when the document is opened by a hosting application.
This setting enables the styles contained in documents
with attached templates to stay synchronized
with the styles used in the attached template.

<the key part is this>
"If the attached template cannot be located or is not a valid file,
then this setting should be silently ignored."

This was added probably in LO 3.6 for a Novell bug for n751020.rtf.

If there is no attachedTemplate provided,
then yes - we need to update styles using Word's application defaults,
but if an attachedTemplate is referenced,
then we need to update the styles using that template.

Since of course LO don't support updating styles from another file
and almost certainly would not have that template available anyway,
just ignore linkStyles when a template is specified.

make CppunitTest_sw_ooxmlexport25 \
    CPPUNIT_TEST_NAME=testTdf166141_linkedStyles

Change-Id: I5ec2f0abeb8073c3caa9a579ba338f30dea9d5ac
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186855
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>