tdf#130857 cui: Don't use multiple names for same radio menu group
In the .ui file for the dialog that can e.g. be triggered
from Writer using "Tools" -> "Language" ->
"More Dictionaries Online...", use the ID of the first
GtkRadioMenuItem of the radio group and set the "group"
property on the other GtkRadioMenuItems belonging to
the same group. Don't set the "group" property on the
menu item that is the first in the group to the ID
of another item of the same group. That isn't necessary
and not handled in QtBuilder.

Change-Id: I954c5aeb97e1175041e4d8cae84d69fe18d8acf1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186862
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins