tdf#130857 qt weld: Run QtInstanceButton::connect_clicked in main thread
Setting a property on a Qt widget needs to happen in
the main thread.

Without this commit, this was seen triggering an assert
in qtbase for the "Tools" -> "Language" ->
"More Dictionaries Online..." dialog in a WIP branch
where support for that dialog is declared for the
SAL_VCL_QT_USE_WELDED_WIDGETS=1 case.

Change-Id: If9e09bed194d9bd34932d353872754b82a2832f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186864
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins