feat(jsdialog): allow editing treeview items
There are some dialogs which we need to edit items in, for example the
bookmarks dialog.

This isn't hooked up to any existing editing functions - the client
won't recieve events that say we intend to edit, say. Instead, the client
should implement its own way to edit an editable item.

This requires the client to know whether a given cell is editable so we
will send that to the client. If the client allows editing non-editable
cells it will cause strange effects (such as editing the document
with text from a different cell in the case of the bookmark dialog).
Additionally, the client needs a way to tell us that it has edited
a cell.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Icd510aab1ce236609e84ed672210d68c4d0c6583
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187193
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>