[API CHANGE]: a11y: Drop css::sheet::AccessibleCsvRuler
Drop the unpublished service
com::sun::star::sheet::AccessibleCsvRuler.

It's not implemented or used anywhere.

ScAccessibleCsvRuler is the class that implements
the functionality mentioned in the IDL file
("The accessible object of the ruler in the CSV import
dialog.") but it doesn't actually implement that service.

A search in Google and on Ask [1] didn't give
any results that would suggest this unpublished service
would be relevant for third party code.

[1] https://ask.libreoffice.org/

Change-Id: I80da59a92ce2e5a38db5230d2af0f6fb4e69d5a4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186909
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>