[API CHANGE]: a11y: Drop css::sheet::AccessibleCsvCell
Drop the unpublished service
com::sun::star::sheet::AccessibleCsvCell.

It's not implemented or used anywhere.

ScAccessibleCsvCell looks like the class implementing
the functionality mentioned ("The accessible object of the data
table in the CSV import dialog") but it doesn't actually
implement that service.

A search in Google and on Ask [1] didn't give
any results that would suggest this unpublished service
would be relevant for third party code.

[1] https://ask.libreoffice.org/

Change-Id: I0abd0bf627ff71ccce148fec1bdfd5c23ba453d4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186911
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>