sw a11y: Let SwAccessibleContext subclass OAccessible
Make SwAccessibleContext an OAccessible subclass.

This allows to reuse logic implemented there
for the XAccessible, XAccessibleContext,
XAccessibleComponent and XAccessibleEventBroadcaster
interfaces.

SwAccessibleContext::getBoundsImpl implements
what is needed to implement OAccessibleComponentHelper::implGetBounds,
so simply call that method from there.

Drop the SolarMutexGuard at the start of
SwAccessibleContext::getBoundsImpl because
the base class takes care of that when calling
OAccessibleComponentHelper::implGetBounds.
Explicitly lock the SolarMutex in
SwAccessibleContext::getLocationOnScreen instead,
which also calls SwAccessibleContext::getBoundsImpl
that doesn't lock the mutex anymore now.

No change in behavior expected or seen in a quick
test with Writer and the gtk3 and qt6 VCL plugins
on Linux together with Accerciser and the Orca screen
reader.

Change-Id: I2b75ccd3eed57f276feab303d63038299ceb90d7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/186988
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins