tdf#126111: Don’t do glyph fallback for PUA codepoints
Private Use Area codepoints are codepoints that Unicode reserves for
private use, so they are typically tied to a specific font and a PUA
code point in a different font can be a totally different symbol, so
applying glyph fallback here is unlikely to give a sensible result.

The bug documents in question misuse PUA codepoints not supported by the
requested font to get an empty box (the font’s .notdef glyph), which
would only happens if no glyph fallabck is done (or no font on the
system supports this PUA code point).

This fixes only the case when requested font is present and we do glyph
fallback. If the requested font is missing, then we might still get PUA
glyphs from the fallback font.

Change-Id: I4688d098b40263e83bc9f2df1833cdc6286b7092
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187012
Tested-by: Jenkins
Reviewed-by: Khaled Hosny <khaled@libreoffice.org>
(cherry picked from commit fcd817175711137c49793018dc9761f13326c863)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187135